requirements.txt
PyYAML
pytest
allure-pytest
requests